requirement.txt
langgraph
langchain
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma
langchain-text-splitters
neo4j
python-dotenv
ipython
jupyter